# Add the option for inactive top cells