Inactive top cells are introduced to add flexibility to the vertical coordinate below ice shelves. 

This PR introduces phase 1 in the development documented in https://github.com/MPAS-Dev/MPAS-Model/pull/811, which touches only the code needed to run a basic ice shelf test case.

The variable `minLevelCell` and related `minLevelEdge*` and `minLevelVertex*` variables are introduced by analogy with `maxLevel*` variables. By default, there are no inactive top cells and `minLevel*` = 1. The main code changes are to k-loop limits and designating surface quantities as those at `minLevel*`.