@xylar 2021-03-10
## Testing

(Edit this comment to add testing as it is performed).

### COMPASS

As a baseline in COMPASS, I'm using the `legacy` branch + https://github.com/xylar/compass/commit/ca592c778ea4144efe4db02c025c8e5ae0be8ca9, which adds a test suite (`z_level_dev`) that includes both all tests in the `nightly` suite and several tests from `land_ice_fluxes`.

The COMPASS branch used for testing is https://github.com/xylar/compass/tree/add_z_level_sub_ice_shelf_2D, but I no COMPASS changes should be required so testing with `legacy` should also work. (I will verify this.)

The results from the `z_level_dev` test suite are bit-for-bit identical to the baseline for:
* Ubuntu 20.04 and Gnu 9.3.0
* Anvil and Intel
* ...

### E3SM

...