@xylar 2021-03-10
@cbegeman, as I noted above, a corresponding PR into `legacy` in the compass repo is apparently going to be required.  I thought `legacy` might just work but apparently some of my changes are needed.