@cbegeman 2021-03-11
> @cbegeman, could you re-test on Grizzly with the new COMPASS branch I created? I'll re-test on Anvil.

Yes, we get bfb with the new COMPASS branch on Grizzly.