@xylar 2021-03-11
@cbegeman, E3SM tests vs. baseline are failing.  Here are some details.

When I do a test merge at the following commits, I am indicating which pass and which fail `SMS.T62_oQU120_ais20.MPAS_LISIO_TEST.anvil_intel`

Commits failing:
* f8819694 Set vertical velocity top to 0 from k=1:minLevelCell-1

Commits passing:
* 73826a31 add minLevel* to ocn_vel_pressure_grad_tend

I will continue updating as new test results come in.