@xylar 2021-03-11
If the QU240 analysis is failing without this, I wonder if it's also failing with ocean/develop. It's hard to see how this could be caused by our changes.