@cbegeman 2021-03-11
So QU240-Analysis is now passing with MPAS-Dev/compass#55, which suggests that the difference is due to the xylar/compass/add_z_level_sub_ice_shelf_2D commits. I'll retest this branch without commit c85f36c with xylar/compass/add_z_level_sub_ice_shelf_2D.