@xylar 2021-03-12
@cbegeman, I haven't had a chance yet but I'll retest everything on Anvil soon, then try again on Compy.