@mattdturner 2021-03-16
@cbegeman Thanks for the work in this PR.  I added a few additional comments to the split explicit driver regarding loop limits.  

Aside from those comments, I think this looks good by visual inspection.  Once those loop limits are updated, I will approve this PR and rely on testing by you and @xylar