@cbegeman 2021-03-16
@mattdturner Thanks for your thorough review! I got those final loop limit changes to split-explicit in.

@xylar and I are working on E3SM tests on Cori, Anvil and Compy so we'll keep you posted on that.