@xylar 2021-03-17
@cbegeman, I'm seeing DIFFs in `SMS.T62_oQU120_ais20.MPAS_LISIO_TEST.compy_pgi`

I'll try rolling back commits to figure out where it's happening.