@xylar 2021-03-17
@cbegeman, I haven't seen this issue in compass testing, but I've never run compass with PGI.  I'm pretty close to having the first commit that causes diffs identified.  I'll let you know.