@xylar 2021-03-17
@cbegeman, some very good news.  I rebased all commits except 73826a3 onto 95daa3f (which was the submodule pointed to by components/mpas-source when I started testing today).  That passed the `SMS.T62_oQU120_ais20.MPAS_LISIO_TEST.compy_pgi` test without any diffs.  So it's just that one commit that I have to debug.