@mark-petersen 2021-03-18
@cbegeman should we give this a final review and shoot for a Monday merge? Are there any unresolved issues?  We just merged https://github.com/MPAS-Dev/MPAS-Model/pull/826 yesterday - there will be a few lines in conflict with this PR. Please rebase on `ocean/develop`.