@xylar 2021-03-21
@cbegeman, as I noted above, I'm getting bit-for-bit results with:
```
SMS_D.T62_oQU120_ais20.MPAS_LISIO_TEST.compy_pgi
```
(that is, with PGI in debug mode) but not when it's optimized.  I think we have to live with the non-bit-for-bit optimized build on Compy because I don't see any way to trick the compiler.  Presumably, it's optimizing something with the old code:
```
do k=2,kMax
```
that can't be optimized with the new code:
```
do k=kMin+1,kMax
```