@cbegeman 2021-03-21
@xylar Thanks for doing that testing. It seems reasonable to me to proceed. 

It looks like commit 8329b9c is the last commit needed after rebasing onto `ocean/develop` after @mattdturner's changes to split-explicit. Did you test with this commit?