@xylar 2021-03-21
> It looks like commit 8329b9c is the last commit needed after rebasing onto ocean/develop after @mattdturner's changes to split-explicit. Did you test with this commit?

I think so but I'm not 100% sure.  Best to just re-test tomorrow.