@cbegeman 2021-03-22
MPAS standalone tests pass (z-level-dev suite, which includes nightly suite).

**The following cases pass on Cori:**
PET_Ln3.T62_oEC60to30v3wLI.GMPAS-DIB-IAF-ISMF.cori-haswell_intel
[To be continued]