@mark-petersen 2021-03-22
@cbegeman and @xylar If those test go well, are you comfortable proceeding? I can make the E3SM PR, and I leave this MPAS PR open for the day.  If everything goes well, we would merge the E3SM PR into `next` today, and make the final merge for both MPAS and E3SM in tomorrow.