@xylar 2021-03-22
@mark-petersen, yes, please go ahead with the E3SM PR.

My tests on Anvil are done (all pass).  My testing on Compy is taking quite a long time.  I will keep you all posted.