@cbegeman 2021-03-22
@mark-petersen I just want to remind you of this PR https://github.com/MPAS-Dev/compass/pull/55 to accompany this one.