@xylar 2021-03-22
I am still working on 3 (slow) tests on Compy.  I will update my "Testing" comment and post the list of completed tests on https://github.com/E3SM-Project/E3SM/pull/4171 once they are done.