@philipwjones 2021-03-22
@mattdturner @cbegeman @xylar Has anyone looked at the performance/timing diffs for these changes in any of the tests with optimization on?