@xylar 2021-03-22
> @mattdturner @cbegeman @xylar Has anyone looked at the performance/timing diffs for these changes in any of the tests with optimization on?

@philipwjones, I don't believe so.  We were not given a clear idea of what we needed to be looking at in which tests.  This is a bit part of what we were requesting your help with both in your review here and in the design doc.