@mattdturner 2021-03-22
@philipwjones I just ran a quick test on Chrysalis.

Test case: `SMS_P2560x1.T62_oRRS18to6v3.GMPAS-IAF.chrysalis_intel`

Before these changes:
```
    OCN Run Time:     698.355 seconds      139.671 seconds/mday         1.69 myears/wday
```

After these changes are applied:
```
    OCN Run Time:     702.525 seconds      140.505 seconds/mday         1.68 myears/wday                    
```