@cbegeman 2021-03-22
@philipwjones Timing differences between the runs appear to be in the noise:

Test case: `SMS_D.T62_oQU120_ais20.MPAS_LISIO_TEST.cori-knl_intel`, 1024 proc.

Before these changes:
`OCN Run Time:    3388.555 seconds      677.711 seconds/mday         0.35 myears/wday`
After these changes:
`OCN Run Time:    3378.931 seconds      675.786 seconds/mday         0.35 myears/wday`

Test case: `PEM_Ln9_P1024.ne30_oECv3.A_WCYCL1850S.cori-knl_intel`

Before these changes:
`OCN Run Time:       5.651 seconds       30.138 seconds/mday         7.85 myears/wday`
After these changes:
`OCN Run Time:       5.618 seconds       29.964 seconds/mday         7.90 myears/wday`